Calling Deactivate isn't necessary on a SingleActiveRegion
Remove code that handles Deactivation as it's duplicating effort as SingleActiveRegion calls deactivate implicitly. Only call Deactivate if we transition to a CurrentPage that is null (say when the last view is removed from the region).
Also, setting the CurrentPage to null when there are actually Children results in undefined behavior. (Mostly the UI locks up)