Call into INavigatedAware when performing region navigation.
When navigating in regions attempt Prism.Forms.Regions IRegionAware (To/From). If its not implemented then attempt Prism.Forms INavigatedAware To/From.  This supports sharing ViewModels between page and region navigation. The If/Else structure is because an IRegionAware implementer could call into INavigatedAware forwarding the INavigationContext.Parameters object and that could end up in 2x call on the INavigatedAware interface.